Move format_code + memoize_on_code to util.code
