+++ docs/util-api.rst
@@ -48,16 +48,25 @@ Beams
 .. autofunction:: africanus.util.beams.beam_filenames
 .. autofunction:: africanus.util.beams.beam_grids
 
+Code
+~~~~
+
+.. currentmodule:: africanus.util.code
+
+.. autosummary::
+    format_code
+    memoize_on_key
+
+.. autofunction:: africanus.util.code.format_code
+.. autoclass:: africanus.util.code.memoize_on_key
+
+
 CUDA
 ~~~~
 
 .. currentmodule:: africanus.util.cuda
 
 .. autosummary::
-    format_kernel
     grids
-    memoize_kernel
 
-.. autofunction:: africanus.util.cuda.format_kernel
 .. autofunction:: africanus.util.cuda.grids
-.. autoclass:: africanus.util.cuda.memoize_kernel
